examples: Move methods vec and tensor from tensor.py to stiffness/utils.py
