+++ devito/types/tensor.py
@@ -80,36 +80,6 @@ class TensorFunction(AbstractTensor):
                                              dimensions=dimensions)
         self._space_dimensions = inds
 
-    @property
-    def vec(self):
-         if self.shape[0] != self.shape[1]:
-             raise Exception("This object is already represented by its vector form.")
-         order = ([(0, 0), (1, 1), (2, 2), (1, 2), (0, 2), (0, 1)]
-             if len(self.space_dimensions) == 3 else [(0, 0), (1, 1), (0, 1)])
-         comp =  [self[o[0],o[1]] for o in order]
-         func = tens_func(self)
-         return func(comp)
-    
-    @property
-    def tensor(self):
-        if self.shape[0] == self.shape[1]:
-            raise Exception("This object is already represented by its tensor form.")
-        ndim = len(self.space_dimensions)
-        M = np.zeros((ndim, ndim), dtype = np.dtype(object))
-        M[0,0] = self[0]
-        M[1,1] = self[1]
-        if len(self.space_dimensions) == 3:
-              M[2,2] = self[2]
-              M[2,1] = self[3]
-              M[1,2] = self[3]
-              M[2,0] = self[4]
-              M[0,2] = self[4]
-        M[1,0] = self[-1]
-        M[0,1] = self[-1]
-
-        func = tens_func(self)
-        return func._new(M)
-
     @classmethod
     def __subfunc_setup__(cls, *args, **kwargs):
         """
